+++ src/Common/ToJsonTrait.php
@@ -11,11 +11,6 @@ trait ToJsonTrait
 {
     public function toJson($options = 0, $depth = 512)
     {
-        // depth was only added on php 5.5
-        if (version_compare(PHP_VERSION, '5.5.0') >= 0) {
-            return json_encode($this->toArray(), $options, $depth);
-        }
-
-        return json_encode($this->toArray(), $options);
+        return json_encode($this->toArray(), $options, $depth);
     }
 }